Multiple markers
In this commit, we extract the latitude and longitude data from a CSV file. This is dont by selecting the appropriate columns in the csv data, and assigning the values underneath them into two lists called lats and lons. Then iterating through both these lists in a for loop, we add markers for each pair of latitude and longitude to the feature group, and create a webMap "map1.html".