Stylized popups
The popups for the markers now show the elevation for the corresponding volcano (data also from the csv file). Similar method used to latitudes and longitudes. There was an error, that popups cannot accept floats, or any non-string data. So changed el to string in the for loop itself. Also added m (meters) for readability.